+++ DB/DBB/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbb/entity/basic/DbT2007KibetsuChoshuYuyoEntity.java
@@ -7,6 +7,7 @@ import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
 import jp.co.ndensan.reams.uz.uza.lang.RDateTime;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
+import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.db.DbTableEntityBase;
 import jp.co.ndensan.reams.uz.uza.util.db.IDbAccessable;
 import jp.co.ndensan.reams.uz.uza.util.db.PrimaryKey;
@@ -36,7 +37,7 @@ public class DbT2007KibetsuChoshuYuyoEntity extends DbTableEntityBase<DbT2007Kib
     @PrimaryKey
     private TsuchishoNo tsuchishoNo;
     @PrimaryKey
-    private int rirekiNo;
+    private Decimal rirekiNo;
     @PrimaryKey
     private RString choshuHoho;
     @PrimaryKey
@@ -148,7 +149,7 @@ public class DbT2007KibetsuChoshuYuyoEntity extends DbTableEntityBase<DbT2007Kib
      *
      * @return 履歴番号
      */
-    public int getRirekiNo() {
+    public Decimal getRirekiNo() {
         return rirekiNo;
     }
 
@@ -157,7 +158,7 @@ public class DbT2007KibetsuChoshuYuyoEntity extends DbTableEntityBase<DbT2007Kib
      *
      * @param rirekiNo 履歴番号
      */
-    public void setRirekiNo(int rirekiNo) {
+    public void setRirekiNo(Decimal rirekiNo) {
         this.rirekiNo = rirekiNo;
     }
 
@@ -237,7 +238,8 @@ public class DbT2007KibetsuChoshuYuyoEntity extends DbTableEntityBase<DbT2007Kib
      * このエンティティの主キーが他の{@literal DbT2007KibetsuChoshuYuyoEntity}と等しいか判定します。
      *
      * @param other 比較するエンティティ
-     * @@return 比較するエンティティが同じ主キーを持つ{@literal DbT2007KibetsuChoshuYuyoEntity}の場合{@literal true}、それ以外の場合は{@literal false}
+     * @@return
+     * 比較するエンティティが同じ主キーを持つ{@literal DbT2007KibetsuChoshuYuyoEntity}の場合{@literal true}、それ以外の場合は{@literal false}
      */
     @Override
     public boolean equalsPrimaryKeys(DbT2007KibetsuChoshuYuyoEntity other) {
