+++ DB/DBB/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbb/entity/basic/DbT2009RentaiGimushaEntity.java
@@ -7,6 +7,7 @@ import jp.co.ndensan.reams.uz.uza.biz.ShikibetsuCode;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.RDateTime;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
+import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.db.DbTableEntityBase;
 import jp.co.ndensan.reams.uz.uza.util.db.IDbAccessable;
 import jp.co.ndensan.reams.uz.uza.util.db.PrimaryKey;
@@ -32,7 +33,7 @@ public class DbT2009RentaiGimushaEntity extends DbTableEntityBase<DbT2009RentaiG
     @PrimaryKey
     private HihokenshaNo hihokenshaNo;
     @PrimaryKey
-    private int rirekiNo;
+    private Decimal rirekiNo;
     private ShikibetsuCode shikibetuCode;
     private FlexibleDate startYMD;
     private FlexibleDate endYMD;
@@ -105,7 +106,7 @@ public class DbT2009RentaiGimushaEntity extends DbTableEntityBase<DbT2009RentaiG
      *
      * @return 履歴番号
      */
-    public int getRirekiNo() {
+    public Decimal getRirekiNo() {
         return rirekiNo;
     }
 
@@ -114,7 +115,7 @@ public class DbT2009RentaiGimushaEntity extends DbTableEntityBase<DbT2009RentaiG
      *
      * @param rirekiNo 履歴番号
      */
-    public void setRirekiNo(int rirekiNo) {
+    public void setRirekiNo(Decimal rirekiNo) {
         this.rirekiNo = rirekiNo;
     }
 
@@ -180,7 +181,8 @@ public class DbT2009RentaiGimushaEntity extends DbTableEntityBase<DbT2009RentaiG
      * このエンティティの主キーが他の{@literal DbT2009RentaiGimushaEntity}と等しいか判定します。
      *
      * @param other 比較するエンティティ
-     * @@return 比較するエンティティが同じ主キーを持つ{@literal DbT2009RentaiGimushaEntity}の場合{@literal true}、それ以外の場合は{@literal false}
+     * @@return
+     * 比較するエンティティが同じ主キーを持つ{@literal DbT2009RentaiGimushaEntity}の場合{@literal true}、それ以外の場合は{@literal false}
      */
     @Override
     public boolean equalsPrimaryKeys(DbT2009RentaiGimushaEntity other) {
